+++ wetterdienst/dwd/util.py
@@ -3,6 +3,7 @@ from typing import Union, Tuple, Optional
 
 import dateparser
 import pandas as pd
+import pytz
 from dateutil.relativedelta import relativedelta
 
 from wetterdienst.dwd.observations.metadata import (
@@ -10,79 +11,25 @@ from wetterdienst.dwd.observations.metadata import (
     DWDObservationResolution,
     DWDObservationPeriod,
 )
-from wetterdienst.dwd.metadata.column_names import DWDMetaColumns
-from wetterdienst.dwd.observations.metadata.column_types import (
-    DATE_FIELDS_REGULAR,
-    DATE_FIELDS_IRREGULAR,
-    QUALITY_FIELDS,
-    INTEGER_FIELDS,
-    STRING_FIELDS,
-)
 from wetterdienst.dwd.metadata.datetime import DatetimeFormat
-from wetterdienst.dwd.observations.metadata.resolution import (
-    RESOLUTION_TO_DATETIME_FORMAT_MAPPING,
-)
 
 
 def build_parameter_set_identifier(
     parameter_set: DWDObservationParameterSet,
     resolution: DWDObservationResolution,
     period: DWDObservationPeriod,
-    station_id: int,
+    station_id: str,
     date_range_string: Optional[str] = None,
 ) -> str:
     """ Create parameter set identifier that is used for storage interactions """
     identifier = (
-        f"{parameter_set.value}/{resolution.value}/"
-        f"{period.value}/station_id_{str(station_id)}"
+        f"{parameter_set.value}/{resolution.value}/" f"{period.value}/{station_id}"
     )
 
     if date_range_string:
         identifier = f"{identifier}/{date_range_string}"
 
     return identifier
-
-
-def coerce_field_types(
-    df: pd.DataFrame, resolution: DWDObservationResolution
-) -> pd.DataFrame:
-    """
-    A function used to create a unique dtype mapping for a given list of column names.
-    This function is needed as we want to ensure the expected dtypes of the returned
-    DataFrame as well as for mapping data after reading it from a stored .h5 file. This
-    is required as we want to store the data in this file with the same format which is
-    a string, thus after reading data back in the dtypes have to be matched.
-
-    Args:
-        df: the station_data gathered in a pandas.DataFrame
-        resolution: time resolution of the data as enumeration
-    Return:
-         station data with converted dtypes
-    """
-
-    for column in df.columns:
-        # Station ids are handled separately as they are expected to not have any nans
-        if column == DWDMetaColumns.STATION_ID.value:
-            df[column] = df[column].astype(int)
-        elif column in DATE_FIELDS_REGULAR:
-            df[column] = pd.to_datetime(
-                df[column],
-                format=RESOLUTION_TO_DATETIME_FORMAT_MAPPING[resolution],
-            )
-        elif column in DATE_FIELDS_IRREGULAR:
-            df[column] = pd.to_datetime(
-                df[column], format=DatetimeFormat.YMDH_COLUMN_M.value
-            )
-        elif column in QUALITY_FIELDS or column in INTEGER_FIELDS:
-            df[column] = pd.to_numeric(df[column], errors="coerce").astype(
-                pd.Int64Dtype()
-            )
-        elif column in STRING_FIELDS:
-            df[column] = df[column].astype(pd.StringDtype())
-        else:
-            df[column] = df[column].astype(float)
-
-    return df
 
 
 def parse_datetime(date_string: str) -> datetime:
@@ -98,7 +45,7 @@ def parse_datetime(date_string: str) -> datetime:
     # Tries out any given format of DatetimeFormat enumeration
     return dateparser.parse(
         date_string, date_formats=[dt_format.value for dt_format in DatetimeFormat]
-    ).replace(tzinfo=None)
+    ).replace(tzinfo=pytz.UTC)
 
 
 def mktimerange(
@@ -135,5 +82,9 @@ def mktimerange(
         raise NotImplementedError(
             "mktimerange only implemented for annual and monthly time ranges"
         )
+
+    # TODO: make tz aware
+    date_from = date_from.replace(tzinfo=pytz.UTC)
+    date_to = date_to.replace(tzinfo=pytz.UTC)
 
     return date_from, date_to